Set read-only permissions to GitHub Actions
This change grants the Actions running these workflows the ability to
read all repository contents, which may not be desired in some cases
depending on the sensitivity of the data in the repository. It's
important to ensure that your GitHub Actions only have the necessary
permissions to perform their intended tasks and no more.