Update GitHub Action dependencies and lint configs
In this commit, I have updated the dependencies for our GitHub Actions by replacing some of them with newer versions. Additionally, I've made modifications to the lint configuration files, updating both yamllint and shfmt in the Makefile. The reason behind these changes is to ensure that our code adheres to the latest best practices and standards.

Here are the specific details of what has been updated:

- Updated GitHub Action dependencies:
  - gaurav-nelson/github-action-markdown-link-check@4a1af151f4d7cf4d8f8ac5780597672a3671b88b to tcort/github-action-markdown-link-check@e7c7a18363c842693fadde5d41a3bd3573a7a225 (version 1.1.2)
  - securego/gosec@bfec35a303e24e2a929b586822eca8486accc670 to securego/gosec@6be2b51fd78feca86af91f5186b7964d76cb1256 (version 2.22.10)
  - reviewdog/action-misspell@e7ea17f144822818706c7e579de7927d59384575 to reviewdog/action-misspell@d6429416b12b09b4e2768307d53bef58d172e962 (version 1.27.0)